+++ e2e-env/.env.dist
@@ -14,3 +14,6 @@ SYNC_API_PORT=3232
 COLLECTOR_API_PORT=3333
 LIVE_SYNC_API_PORT=3434
 WS_PORT=8080
+
+UID=1000
+GID=1000
