+++ e2e-env/cloudsync-mock/src/lib/server.ts
@@ -9,20 +9,22 @@ export class Server {
 
   public constructor(port: number) {
     this.api = express();
-    
-    this.api.use(this.middleware.bind(this));
-    this.port = port;
 
-    this.wsServer = WsServer.getInstance();
+    const wsServer = WsServer.getInstance();
+
+    this.api.use((req: Request, res: Response, next: NextFunction) => {
+     wsServer.sendDataToWS(this.constructor.name, req);
+      next();
+    });
+    this.api.use((req: Request, res: Response, next: NextFunction) => {
+      wsServer.sendDataToWS(this.constructor.name, req);
+      next();
+    });
+    this.port = port;
   }
 
   public async listen() {
     console.log(`${this.constructor.name} listening on port \x1b[96m${this.port}\x1b[0m`);
     return this.api.listen(this.port);
-  }
-
-  private middleware(req: Request, res: Response, next: NextFunction) {
-    this.wsServer.sendDataToWS(this.constructor.name, req);
-    next();
   }
 }
